Remove `StudentViewer` object and inline its functionality into `Functions` to simplify code and reduce redundancy.
