--- src/main/kotlin/StudentViewer.kt
@@ -1,21 +1,0 @@
-package org.example
-
-object StudentViewer {
-    fun displayAll(students: List<Student>) {
-        if (students.isEmpty()) {
-            println("No students found.")
-            return
-        }
-
-        println("\n=== STUDENT LIST ===")
-        students.forEach { student ->
-            println("ID: ${student.id}")
-            println("Name: ${student.name}")
-            println("GPA: ${student.gpa ?: "N/A"}")
-            println("Grade: ${student.grade ?: "N/A"}")
-            println("Status: ${student.status ?: "N/A"}")
-            println("Notes: ${if (student.notes.isNullOrEmpty()) "No notes" else student.notes?.joinToString(", ")}")
-            println("-------------------")
-        }
-    }
-}
